started to add istio verify-install command
